refactor(app): extract view/camera and model-lookup modules
Move camera/view update, collision, terrain snapping, and marker manipulation (ProcessInput, UpdateViewerVectors/ViewDefine, EditorProcessClick, CheckCollision, SnapObjectsToTerrain, UpdateMarkerManipulation) into app_view.cpp; and model name/id lookup/copy/search plus validation (Lookup*/Copy*/SearchModel*, GetVisibleTreeNodes, IsComputer/IsWaterTower, ValidateParentChildCompatibility, EvaluateTrainTrackPositions) with their containsIgnoreCase/ModelEntry/LoadAllModelsFromJson helpers into app_lookup.cpp. app.cpp core 5791 -> 2447 lines. Pure code motion: builds clean, runtime-verified (level 2, 558 objects, no errors).
